# uStoreAPI/uStoreAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "uStoreAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Calificaciones/GetCalificacionesTienda": {
      "get": {
        "tags": [
          "Calificaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CalificacionTiendaDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/GetCalificacionesProducto": {
      "get": {
        "tags": [
          "Calificaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idProducto",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CalificacionProductoDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/GetCalificacionTienda": {
      "get": {
        "tags": [
          "Calificaciones"
        ],
        "operationId": "GetCalificacionTienda",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CalificacionTiendaDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/GetCalificacionProducto": {
      "get": {
        "tags": [
          "Calificaciones"
        ],
        "operationId": "GetCalificacionProducto",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CalificacionProductoDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/CreateCalificacionTienda": {
      "post": {
        "tags": [
          "Calificaciones"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CalificacionTiendaDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CalificacionTiendum"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/CreateCalificacionProducto": {
      "post": {
        "tags": [
          "Calificaciones"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CalificacionProducto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CalificacionProducto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/UpdateCalificacionTienda": {
      "put": {
        "tags": [
          "Calificaciones"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CalificacionTiendaDto"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/UpdateCalificacionProducto": {
      "put": {
        "tags": [
          "Calificaciones"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CalificacionProductoDto"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/DeleteCalificacionTienda": {
      "delete": {
        "tags": [
          "Calificaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Calificaciones/DeleteCalificacionProducto": {
      "delete": {
        "tags": [
          "Calificaciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CalificacionProducto": {
      "type": "object",
      "properties": {
        "idCalificacionProducto": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idProductosNavigation": {
          "$ref": "#/definitions/Producto"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
    "CalificacionProductoDto": {
      "required": [
        "calificacion",
        "idProductos"
      ],
      "type": "object",
      "properties": {
        "idCalificacionProducto": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "CalificacionTiendaDto": {
      "required": [
        "calificacion",
        "idTienda"
      ],
      "type": "object",
      "properties": {
        "idCalificacionTienda": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "CalificacionTiendum": {
      "type": "object",
      "properties": {
        "idCalificacionTienda": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "CategoriasTienda": {
      "type": "object",
      "properties": {
        "idCt": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idCategoria": {
          "format": "int32",
          "type": "integer"
        },
        "idCategoriaNavigation": {
          "$ref": "#/definitions/Categoria"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        }
      },
      "additionalProperties": false
    },
...
    "Comentario": {
      "type": "object",
      "properties": {
        "idComentarios": {
          "format": "int32",
          "type": "integer"
        },
        "comentario1": {
          "type": "string"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idProductosNavigation": {
          "$ref": "#/definitions/Producto"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "Favorito": {
      "type": "object",
      "properties": {
        "idFavoritos": {
          "format": "int32",
          "type": "integer"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idProductosNavigation": {
          "$ref": "#/definitions/Producto"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "Producto": {
      "type": "object",
      "properties": {
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProducto": {
          "type": "string"
        },
        "precioProducto": {
          "format": "double",
          "type": "number"
        },
        "cantidadApartado": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "alertaApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlertaApartado"
          }
        },
        "calificacionProductos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalificacionProducto"
          }
        },
        "categoriasProductos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CategoriasProducto"
          }
        },
        "comentarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comentario"
          }
        },
        "favoritos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Favorito"
          }
        },
        "historials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historial"
          }
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "imagenesProductos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImagenesProducto"
          }
        },
        "solicitudesApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SolicitudesApartado"
          }
        },
        "tendenciasVenta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TendenciasVentum"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Tiendum": {
      "type": "object",
      "properties": {
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTienda": {
          "type": "string"
        },
        "rangoPrecio": {
          "type": "string"
        },
        "idCentroComercial": {
          "format": "int32",
          "type": "integer"
        },
        "idAdministrador": {
          "format": "int32",
          "type": "integer"
        },
        "apartados": {
          "format": "int32",
          "type": "integer"
        },
        "vistas": {
          "format": "int32",
          "type": "integer"
        },
        "logoTienda": {
          "type": "string"
        },
        "calificacionTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalificacionTiendum"
          }
        },
        "categoriasTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CategoriasTienda"
          }
        },
        "comentarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comentario"
          }
        },
        "favoritos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Favorito"
          }
        },
        "gerentes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Gerente"
          }
        },
        "historials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historial"
          }
        },
        "horarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Horario"
          }
        },
        "idAdministradorNavigation": {
          "$ref": "#/definitions/AdministradorTiendum"
        },
        "idCentroComercialNavigation": {
          "$ref": "#/definitions/CentroComercial"
        },
        "imagenesTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImagenesTienda"
          }
        },
        "periodosPredeterminados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PeriodosPredeterminado"
          }
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Producto"
          }
        },
        "publicaciones": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Publicacione"
          }
        },
        "solicitudesApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SolicitudesApartado"
          }
        },
        "tendenciasVenta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TendenciasVentum"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Usuario": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idDetallesUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "alertaApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AlertaApartado"
          }
        },
        "calificacionProductos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalificacionProducto"
          }
        },
        "calificacionTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalificacionTiendum"
          }
        },
        "comentarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comentario"
          }
        },
        "cuentaUsuarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CuentaUsuario"
          }
        },
        "favoritos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Favorito"
          }
        },
        "historials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historial"
          }
        },
        "idDetallesUsuarioNavigation": {
          "$ref": "#/definitions/DetallesUsuario"
        },
        "solicitudesApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SolicitudesApartado"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}